+++ packages/grid/src/components/cell-editors/abstract-cell-editor.component.ts
@@ -1,7 +1,8 @@
-import { ChangeDetectionStrategy, Component, computed, ElementRef, inject, Input, OnInit, output } from '@angular/core';
+import { ChangeDetectionStrategy, Component, computed, ElementRef, inject, input, Input, OnInit, output } from '@angular/core';
 import { ThyPopoverRef } from 'ngx-tethys/popover';
 import { AITable, AITableQueries } from '../../core';
 import { AITableField, AITableReferences, UpdateFieldValueOptions } from '@ai-table/utils';
+import { AI_TABLE_RECORD_HEIGHT_LEVELS } from '../../constants';
 
 @Component({
     selector: 'abstract-edit-cell',
@@ -16,6 +17,8 @@ export abstract class AbstractEditCellEditor<TValue, TFieldType extends AITableF
     @Input({ required: true }) recordId!: string;
 
     @Input({ required: true }) references!: AITableReferences;
+
+    recordHeight = input<number>(AI_TABLE_RECORD_HEIGHT_LEVELS.low);
 
     updateFieldValues = output<UpdateFieldValueOptions<TValue>[]>();