+++ packages/grid/src/services/event.service.ts
@@ -164,7 +164,8 @@ export class AITableGridEventService {
                 recordId: recordId,
                 references,
                 aiTable: aiTable,
-                isSelectAll: options.isSelectAll
+                isSelectAll: options.isSelectAll,
+                rowHeight: offsetOriginPosition.height
             },
             panelClass: 'grid-cell-editor',
             outsideClosable: fieldType === AITableFieldType.link ? true : false,