+++ packages/grid/src/types/canvas.ts
@@ -67,10 +67,13 @@ export interface AITableText extends AITableGraph {
     verticalAlign?: 'top' | 'middle' | 'bottom';
     fontWeight?: AITableFontWeight;
     textDecoration?: 'underline' | 'line-through' | 'none';
+    wrap?: 'char' | 'word' | 'none';
+    ellipsis?: boolean;
 }
 
 export interface AITableWrapText extends AITableText {
     maxWidth: number;
+    maxHeight?: number;
     lineHeight: number;
     fieldType: AITableFieldType | string;
     maxRow?: number;