+++ src/app/examples/basic/basic.component.ts
@@ -2,12 +2,11 @@ import { Component, signal } from '@angular/core';
 import { AITableGrid } from '@ai-table/grid';
 import { AITableRecord, AITableField, AITableReferences } from '@ai-table/utils';
 import { mockRecords, mockFields, mockReferences } from './mock';
-import { ThyPopoverModule } from 'ngx-tethys/popover';
 
 @Component({
-    selector: 'app-basic-table-example',
+    selector: 'app-table-basic-example',
     templateUrl: './basic.component.html',
-    imports: [AITableGrid, ThyPopoverModule],
+    imports: [AITableGrid],
     host: {
         class: 'd-block w-100 h-100'
     }