+++ src/app/examples/custom-field/relation/index.ts
@@ -2,3 +2,4 @@ export * from './types';
 export * from './field-model';
 export * from './render';
 export * from './cover-render';
+export * from './types';